Μεταφορά login στην αρχική οθόνη και ανανέωση UI
