+++ app/src/main/java/com/ioannapergamali/mysmartroute/view/ui/screens/AnnounceTransportScreen.kt
@@ -1,8 +1,6 @@
 package com.ioannapergamali.mysmartroute.view.ui.screens
 
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -22,6 +20,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.res.stringResource
+import com.ioannapergamali.mysmartroute.view.ui.components.ScreenContainer
 import androidx.compose.foundation.layout.Row
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
@@ -177,7 +176,7 @@ fun AnnounceTransportScreen(navController: NavController, openDrawer: () -> Unit
         endLatLng?.let { toMarkerState.position = it }
     }
 
-    Column(modifier = Modifier.fillMaxSize().padding(16.dp)) {
+    ScreenContainer(modifier = Modifier.padding(0.dp)) {
         TopBar(
             title = "Announce Transport",
             navController = navController,
@@ -596,7 +595,6 @@ fun AnnounceTransportScreen(navController: NavController, openDrawer: () -> Unit
         if (state is TransportAnnouncementViewModel.AnnouncementState.Error) {
             Spacer(modifier = Modifier.height(8.dp))
             Text(text = (state as TransportAnnouncementViewModel.AnnouncementState.Error).message)
-        }
     }
 
     LaunchedEffect(state) {
