+++ app/src/main/java/com/ioannapergamali/mysmartroute/view/ui/screens/DirectionsMapScreen.kt
@@ -4,10 +4,13 @@ import android.webkit.WebView
 import android.webkit.WebViewClient
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.border
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.navigation.NavController
 import com.ioannapergamali.mysmartroute.view.ui.components.TopBar
@@ -41,6 +44,7 @@ fun DirectionsMapScreen(
             modifier = Modifier
                 .fillMaxSize()
                 .padding(paddingValues)
+                .border(2.dp, MaterialTheme.colorScheme.primary)
         )
     }
 }
