+++ app/src/main/java/com/ioannapergamali/mysmartroute/view/ui/screens/LocalDatabaseScreen.kt
@@ -1,8 +1,6 @@
 package com.ioannapergamali.mysmartroute.view.ui.screens
 
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -20,6 +18,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.ioannapergamali.mysmartroute.view.ui.components.TopBar
+import com.ioannapergamali.mysmartroute.view.ui.components.ScreenContainer
 import com.ioannapergamali.mysmartroute.viewmodel.DatabaseViewModel
 
 @Composable
@@ -41,17 +40,12 @@ fun LocalDatabaseScreen(navController: NavController, openDrawer: () -> Unit) {
         }
     ) { padding ->
         if (data == null) {
-            Column(modifier = Modifier.fillMaxSize().padding(padding),
-                ) {
+            ScreenContainer(modifier = Modifier.padding(padding)) {
                 CircularProgressIndicator(modifier = Modifier.padding(16.dp))
             }
         } else {
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(padding)
-                    .padding(16.dp)
-            ) {
+            ScreenContainer(modifier = Modifier.padding(padding)) {
+                LazyColumn(modifier = Modifier.fillMaxSize()) {
                 item { Text("Users", style = MaterialTheme.typography.titleMedium) }
                 if (data!!.users.isEmpty()) {
                     item { Text("Ο πίνακας είναι άδειος") }
@@ -86,6 +80,7 @@ fun LocalDatabaseScreen(navController: NavController, openDrawer: () -> Unit) {
                     items(data!!.settings) { setting ->
                         Text("${setting.userId} - ${setting.theme}")
                     }
+                }
                 }
             }
         }
