+++ app/src/main/java/com/ioannapergamali/mysmartroute/view/ui/screens/SupportScreen.kt
@@ -1,13 +1,12 @@
 package com.ioannapergamali.mysmartroute.view.ui.screens
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.ioannapergamali.mysmartroute.view.ui.components.ScreenContainer
 import androidx.navigation.NavController
 import com.ioannapergamali.mysmartroute.view.ui.components.TopBar
 
@@ -23,7 +22,7 @@ fun SupportScreen(navController: NavController, openDrawer: () -> Unit) {
             )
         }
     ) { padding ->
-        Column(modifier = Modifier.fillMaxSize().padding(padding).padding(16.dp)) {
+        ScreenContainer(modifier = Modifier.padding(padding)) {
             Text("Email: smartroute@info.gr")
             Text("Address: Πάροδος Κρήτης 8, Γάζι, ΤΚ 71414")
         }
